aidabackend/requirements.txt
asgiref==3.8.1
contourpy==1.3.0
cycler==0.12.1
cytoolz==0.12.3
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
eth-hash==0.7.0
eth-keys==0.5.1
eth-typing==5.0.0
eth-utils==5.0.0
filelock==3.16.0
fonttools==4.54.1
hexbytes==1.2.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.3
psycopg2==2.9.9
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.4
wheel==0.44.0
